feat: 편집 화면의 드롭다운을 항상 펼쳐진 칩으로 바꾼다 (v0.2.2)
가계부·여행 앱 벤치마킹 결과를 반영한다. 지출 카테고리 같은 고정 소수 선택지를
드롭다운으로 받는 것은 "펼치기 → 고르기" 두 탭이고, 펼치기 전에는 선택지가
보이지 않는다. 가계부 앱들은 카테고리를 타일·칩으로 항상 펼쳐 둔다 — 한 탭으로
끝나고 현재 값과 대안이 같이 보인다.

## ChipSelector (신규 공용 컴포넌트)

- FlowRow 줄바꿈. 가로 스크롤 한 줄이 아니다 — 화면 밖으로 밀린 칩은 없는 것과
  같아서 "전체가 한눈에"라는 목적이 깨진다.
- 교체 기준을 컴포넌트 주석에 남겼다: 고정이고 10개 미만이면 칩, 도시·통화·연결
  일정처럼 길거나 동적이면 드롭다운 유지(칩 20개는 화면 절반을 먹는다).

교체한 곳 8군데:
- 지출: 카테고리(7)·결제자(미지정+2인)
- 예약: 유형(6)·상태(6)
- 장소: 카테고리(6)·우선순위(3)
- 일정: 유형(6)·담당자(없음+2인)

유지한 곳: 통화(18)·도시(동적)·시간대(라벨 긺)·연결 일정/예약(동적).

## 지출 날짜 "오늘 / 어제" 칩

여행 중에는 어제 쓴 돈을 오늘 몰아서 적는 일이 잦다. 달력 다이얼로그(탭 3번)
대신 칩 한 탭. 그 외 날짜는 기존 달력 그대로.

## 되돌리기 4초 → 10초 (SnackbarDuration.Long)

v0.2.1 실기기 검증에서 사람이 스낵바를 읽고 → 손을 옮겨 → 누르기 전에 사라지는
일이 실제로 났다. 삭제 직후는 "지워졌나?"를 확인하는 시간이 먼저라 4초는 짧다.

## 실기기 검증 (R5KL20G4NWJ, 0.2.2)

- 지출 편집: 카테고리 7칩 줄바꿈 렌더, "교통" 한 탭 → checked=true 전환 확인
- 결제자 칩(미지정·이정민), 오늘/어제 칩(오늘 선택 상태) 렌더 확인 (스크린샷)
- 예약 편집: 유형 6칩·상태 6칩 렌더 확인 (스크린샷)

+++ app/src/main/kotlin/com/jeongmin/honeymoondoctor/core/ui/UndoDeleteSnackbar.kt
@@ -29,7 +29,10 @@ fun UndoDeleteSnackbarEffect(
         val result = hostState.showSnackbar(
             message = pending.message,
             actionLabel = "되돌리기",
-            duration = SnackbarDuration.Short,
+            // Long(10초). 처음엔 Short(4초)였는데 실기기 검증에서 사람이 스낵바를
+            // 읽고 → 손을 옮겨 → 누르기 전에 사라지는 일이 실제로 났다. 삭제 직후는
+            // "지워졌나?"를 확인하는 시간이 먼저라 4초는 짧다.
+            duration = SnackbarDuration.Long,
         )
         when (result) {
             SnackbarResult.ActionPerformed -> onUndo()
